[Pallas] Move runtime/compact_worklist.py into runtime/pallas
stack-info: PR: https://github.com/pytorch/helion/pull/3384, branch: AmesingFlank/stack/115
